requirements.txt
# Core Dependencies
flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
ollama>=0.1.6
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-flask>=1.3.0
